@@ -22,7 +22,6 @@ import android.media.AudioManager;
 import android.os.Build;
 import android.os.SystemClock;
 import android.provider.Settings;
-import android.support.v4.view.accessibility.AccessibilityEventCompat;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.MotionEvent;
@@ -37,6 +36,8 @@ import com.onyx.latinime.R;
 import com.onyx.latinime.compat.SettingsSecureCompatUtils;
 import com.onyx.latinime.latin.SuggestedWords;
 import com.onyx.latinime.latin.utils.InputTypeUtils;
+
+import androidx.core.view.accessibility.AccessibilityEventCompat;
 
 public final class AccessibilityUtils {
     private static final String TAG = AccessibilityUtils.class.getSimpleName();
