@@ -111,8 +111,7 @@ final public class BinaryDictionaryGetter {
         final SharedPreferences mDictPreferences;
         public DictPackSettings(final Context context) {
             mDictPreferences = null == context ? null
-                    : context.getSharedPreferences(COMMON_PREFERENCES_NAME,
-                            Context.MODE_WORLD_READABLE | Context.MODE_MULTI_PROCESS);
+                    : context.getSharedPreferences(COMMON_PREFERENCES_NAME, Context.MODE_PRIVATE);
         }
         public boolean isWordListActive(final String dictId) {
             if (null == mDictPreferences) {
