@@ -1,8 +1,9 @@
 package com.onyx.calculator;
 
 import android.content.Context;
-import android.support.v7.widget.LinearLayoutManager;
 import android.util.AttributeSet;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
 
 /**
  * Created by ming on 16/9/14.
