@@ -17,11 +17,12 @@
 package com.onyx.deskclock.deskclock;
 
 import android.app.Activity;
-import android.app.Fragment;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageButton;
 import android.widget.ImageView;
+
+import androidx.fragment.app.Fragment;
 
 public class DeskClockFragment extends Fragment {
 
