@@ -26,13 +26,14 @@ import android.database.ContentObserver;
 import android.os.BatteryManager;
 import android.os.Handler;
 import android.provider.Settings;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.view.Window;
 import android.view.WindowManager;
 
 import com.onyx.deskclock.R;
 import com.onyx.deskclock.deskclock.widget.TextTime;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class ScreensaverActivity extends AppCompatActivity {
     static final boolean DEBUG = false;
