@@ -17,9 +17,8 @@
 package com.onyx.deskclock.deskclock.alarms.dataadapter;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SwitchCompat;
 import android.view.View;
+import android.widget.CheckBox;
 import android.widget.CompoundButton;
 import android.widget.TextView;
 
@@ -29,6 +28,8 @@ import com.onyx.deskclock.deskclock.alarms.AlarmTimeClickHandler;
 import com.onyx.deskclock.deskclock.provider.Alarm;
 import com.onyx.deskclock.deskclock.provider.AlarmInstance;
 import com.onyx.deskclock.deskclock.widget.TextTime;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Abstract ViewHolder for alarm time items.
@@ -89,8 +90,7 @@ public abstract class AlarmTimeViewHolder extends RecyclerView.ViewHolder {
 
     protected void bindOnOffSwitch(Context context, Alarm alarm) {
         onoff.setChecked(alarm.enabled);
-        ((SwitchCompat) onoff).setTextOn(context.getString(R.string.on_switch));
-        ((SwitchCompat) onoff).setTextOff(context.getString(R.string.off_switch));
+        ((CheckBox) onoff).setContentDescription(alarm.enabled ? context.getString(R.string.on_switch) : context.getString(R.string.off_switch));
     }
 
     protected void bindClock(Context context, Alarm alarm) {
