@@ -16,7 +16,7 @@
 
 package com.onyx.deskclock.deskclock.widget.toast;
 
-import android.support.design.widget.Snackbar;
+import com.google.android.material.snackbar.Snackbar;
 
 import java.lang.ref.WeakReference;
 
